+++ src/main/java/pro/sky/telegrambot/model/NotificationTask.java
@@ -8,14 +8,10 @@ import java.util.Objects;
 @Table(name = "notification_task")
 public class NotificationTask {
     @Id
-//    @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-//    @Column(name = "chat")
     private Long chat;
-//    @Column(name = "message")
     private String message;
-//    @Column(name = "send_date_time")
     private LocalDateTime send_date_time;
 
     public NotificationTask(Long id, Long chat, String message, LocalDateTime send_date_time) {
