+++ src/main/resources/application.properties
@@ -1,0 +1,8 @@
+spring.datasource.url=jdbc:postgresql://localhost:5432/TelegramBasa
+spring.datasource.username=MyUser
+spring.datasource.password=456338
+telegram.bot.token=6503866298:AAGHmX4pkGj2vUbr3evX4ZlgDe1IeY2xyY8
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
+
+spring.liquibase.change-log=classpath:liquibase/changelot-master.yml
